pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-eval-with-pytest-ragas"
version = "0.1.0"
description = "Evaluate LLM applications using pytest + RAGAS — PyCon HK 2026 demo"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "unrealandychan", email = "unrealandychan@users.noreply.github.com" },
]
keywords = ["llm", "eval", "ragas", "pytest", "rag", "ai", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "ragas>=0.2",
    "openai>=1.0",
    "anthropic>=0.20",
    "langchain-openai>=0.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "datasets>=2.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/unrealandychan/llm-eval-with-pytest-ragas"
Repository = "https://github.com/unrealandychan/llm-eval-with-pytest-ragas"
Issues = "https://github.com/unrealandychan/llm-eval-with-pytest-ragas/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_eval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "llm: marks tests that call real LLM APIs (deselect with '-m not llm')",
    "ragas: marks tests that use RAGAS evaluation framework",
    "integration: marks end-to-end integration tests",
    "slow: marks tests that are slow to run",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
